feat(oceanbase-vector-store): align implementation with Spring AI VectorStore contract
Fix compatibility issues with BaseVectorStoreTests.

- switch primary key to VARCHAR(36) to support String-based document IDs
- implement upsert semantics for add() operations
- use cosine distance as the default metric
- add similarityThreshold filtering for search()
- convert distance values to normalized similarity scores
